Move client-side message handler into a separate package (#1462)
* Move the client-side message handler into a separate package
* Update package docstring
