+++ src/py/flwr/client/message_handler/message_handler_test.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Client message handler tests."""
+"""Client-side message handler tests."""
 
 
 from flwr.client import Client
