+++ src/py/flwr/server/grpc_server/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Server-side networking package."""
+"""Server-side part of the gRPC transport layer."""
